added jsdoc comments
