+++ packages/core/src/cache/containsGlobalDeclarations.ts
@@ -1,0 +1,39 @@
+// eslint-disable-next-line n/no-extraneous-import
+import ts from "typescript";
+
+/**
+ * Inspects top-level statements of a TS source file to determine
+ * if it introduces or modifies entities in the global scope.
+ * @param rawFileContent The raw source code of the file to analyze.
+ * @returns `true` if the file contains explicit global augmentations
+ * (e.g., `declare global {}`) or top-level ambient declarations (e.g., `declare const`),
+ * otherwise `false`.
+ */
+export function containsGlobalDeclarations(rawFileContent: string) {
+	const sourceFileNode = ts.createSourceFile(
+		"mayContainGlobals.ts",
+		rawFileContent,
+		ts.ScriptTarget.ESNext,
+		true,
+	);
+
+	return sourceFileNode.statements.some((statement) => {
+		// checks for 'declare global {}'
+		if (ts.isModuleDeclaration(statement) && statement.name.text === "global") {
+			return true;
+		}
+
+		// checks for 'declare' keyword
+		const canHaveModifiers = ts.canHaveModifiers(statement);
+		if (!canHaveModifiers) {
+			return false;
+		}
+
+		const modifiers = ts.getModifiers(statement);
+		const hasDeclareKeyword = modifiers?.some(
+			(mod) => mod.kind === ts.SyntaxKind.DeclareKeyword,
+		);
+
+		return hasDeclareKeyword;
+	});
+}
